app/databases/init/requirements.txt
requests
qdrant_client
nanoid
numpy
pydantic
pandas
pillow